feat: Start Session Behaviors
- All previous changes in the WIP
- Now shows current number of player in start page
- Added test for `AddElements` in `ElementDataProvider`
- Added FIX: `AddElements` to use instead of a foreach
